refactor: move html conversion functions
