+++ src/html/thematic_break.ts
@@ -1,0 +1,7 @@
+import type { TsmarkNode } from '../types.d.ts';
+
+export function thematicBreakToHTML(
+  _node: TsmarkNode & { type: 'thematic_break' },
+): string {
+  return '<hr />';
+}
